+++ src/entities/Article/ui/ArticleDetails/ArticleDetails.module.scss
@@ -3,92 +3,3 @@
   justify-content: center;
   align-items: center;
 }
-
-.articleBlock {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  padding: 0px 40px;
-  width: 100%;
-}
-
-.articleBlockHeader {
-  display: flex;
-  flex-direction: column;
-  margin-top: 30px;
-  width: 100%;
-}
-
-.articleBlockHeaderTop {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  flex-direction: column;
-  width: 100%;
-}
-
-.articleBlockHeaderTopAvatar {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.articleBlockTopTitle {
-  margin-left: 10px;
-}
-
-.articleBlockHeaderArticleSubtitle {
-  display: flex;
-  justify-content: flex-start;
-  align-items: flex-start;
-  width: 80%;
-  margin-top: 10px;
-}
-
-.articleHeaderTopStat {
-  display: flex;
-  flex-direction: row;
-  margin-top: 15px;
-  width: 35%;
-}
-
-.articleHeaderTopStatViews {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  width: 50%;
-}
-
-.articleHeaderTopStatData {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  width: 80%;
-}
-
-.viewsIcon {
-  width: 25px;
-  height: 25px;
-  margin-right: 5px;
-}
-
-.dateIcon {
-  width: 20px;
-  height: 20px;
-  margin-right: 5px;
-}
-
-.articleContent {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 50px;
-  width: 100%;
-  height: 100%;
-  margin-top: 50px;
-}
-
-.articleContentBlock {
-  margin-top: 10px;
-}
