+++ src/entities/Article/ui/ArticleDetails/ArticleDetails.tsx
@@ -29,6 +29,9 @@ import {
   DoubleAdjustableBlock,
 } from 'shared/ui/Block/DoubleAdjustableBlock/DoubleAdjustableBlock';
 import {
+  ArticleDetailsContent,
+} from 'entities/Article/ui/ArticleDetailsContent/ArticleDetailsContent';
+import {
   ArticleImageBlockComponent,
 } from '../ArticleImageBlockComponent/ArticleImageBlockComponent';
 import { ArticleCodeBlockComponent } from '../ArticleCodeBlockComponent/ArticleCodeBlockComponent';
@@ -58,23 +61,9 @@ export const ArticleDetails = memo((props: ArticleDetailsProps) => {
     id,
   } = props;
   const { t } = useTranslation('articles');
-  const data = useSelector(getArticleDetails);
   const isLoading = useSelector(getArticleIsLoading);
   const error = useSelector(getArticleError);
   const dispatch = useAppDispatch();
-
-  const renderBlock = useCallback((block: ArticleBlock) => {
-    switch (block.type) {
-      case ArticleBlockType.TEXT:
-        return <ArticleTextBlockComponent block={block} />;
-      case ArticleBlockType.CODE:
-        return <ArticleCodeBlockComponent />;
-      case ArticleBlockType.IMAGE:
-        return <ArticleImageBlockComponent />;
-      default:
-        return null;
-    }
-  }, []);
 
   useEffect(() => {
     dispatch(fetchArticleById(id));
@@ -99,85 +88,7 @@ export const ArticleDetails = memo((props: ArticleDetailsProps) => {
     );
   } else {
     content = (
-        <DoubleAdjustableBlock
-            children={(
-                <div className={classes.articleBlock}>
-                    <div className={classes.articleBlockHeader}>
-                        <div className={classes.articleBlockHeaderTop}>
-                            <div className={classes.articleBlockHeaderTopAvatar}>
-                                <Avatar
-                                    size={AvatarSize.L}
-                                    src={data?.img}
-                                    alt={data?.title}
-                                />
-                                <Text
-                                    className={classes.articleBlockTopTitle}
-                                    theme={TextTheme.BLOCK_TEXT}
-                                    title={t(data?.title ? data?.title : `Статья #${data?.id}`)}
-                                    align={TextAlign.LEFT}
-                                    size={TextSize.L}
-                                />
-                            </div>
-                            {
-                    data?.subtitle
-                      ? (
-                          <div className={classes.articleBlockHeaderArticleSubtitle}>
-                              <Text
-                                  theme={TextTheme.SUBTITLE}
-                                  text={t(data?.subtitle)}
-                                  align={TextAlign.LEFT}
-                                  size={TextSize.S}
-                              />
-                          </div>
-                      )
-                      : (
-                          <Text
-                              theme={TextTheme.SUBTITLE}
-                              text="-"
-                              align={TextAlign.LEFT}
-                              size={TextSize.S}
-                          />
-                      )
-                  }
-                            <div className={classes.articleHeaderTopStat}>
-                                <div className={classes.articleHeaderTopStatViews}>
-                                    <Icon
-                                        className={classes.viewsIcon}
-                                        Svg={ViewsIcon}
-                                        theme={IconTheme.BLOCK_ICON}
-                                    />
-                                    <Text
-                                        theme={TextTheme.BLOCK_TEXT}
-                                        text={t(String(data?.views))}
-                                        align={TextAlign.LEFT}
-                                        size={TextSize.S}
-                                    />
-                                </div>
-                                <div className={classes.articleHeaderTopStatData}>
-                                    <Icon
-                                        className={classes.dateIcon}
-                                        Svg={DateIcon}
-                                        theme={IconTheme.BLOCK_ICON}
-                                    />
-                                    <Text
-                                        theme={TextTheme.BLOCK_TEXT}
-                                        text={t(String(data?.createdAt))}
-                                        align={TextAlign.LEFT}
-                                        size={TextSize.S}
-                                    />
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                    <div className={classes.articleContent}>
-                        <div className={classes.articleContentBlock}>
-                            {data?.blocks.map(renderBlock)}
-                        </div>
-                    </div>
-                </div>
-              )}
-            recommendations={<RecommendationsBlock>hello</RecommendationsBlock>}
-        />
+        <ArticleDetailsContent id={id} />
     );
   }
